requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
defusedxml==0.6.0
Django==3.0
django-allauth==0.41.0
django-braces==1.13.0
django-cors-headers==3.2.0
django-filter==2.2.0
django-oauth-toolkit==1.2.0
django-rest-auth==0.9.5
django-rest-framework-social-oauth2==1.1.0
django-templated-mail==1.1.1
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
httpie==1.0.3
idna==2.8
Markdown==3.1.1
oauth2-provider==0.0
oauthlib==3.1.0
Pillow==6.2.1
psycopg2==2.8.4
Pygments==2.5.2
PyJWT==1.7.1
python-decouple==3.3
python-social-auth==0.3.6
python3-openid==3.1.0
pytz==2019.3
PyYAML==5.2
requests==2.22.0
requests-oauthlib==1.3.0
rest-django-jwt==0.0.1
six==1.13.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sqlparse==0.3.0
urllib3==1.25.7
